debugging loading balances for wallet
when we listen for records from live query they come in descending order. so when we quickly click on wallet the balance is set for the last thing that is created in db which is the oldest balance. check i added prevents balance being overwritten by some older historical balance.